add "bundle" target that will create a tarball for uploading
It builds and tests the project in cwd to reveal its dependencies,
tries to get its hands on the distribution name and description,
and proposes a META.info file with this just collected information.

Sadly the used prompt() segfaults on moar and besides that we need
to autodiscover the types it is going to provide, similar to the
dependency check.